backend/requirements.txt
# FILE: requirements.txt
fastapi
uvicorn[standard]
anthropic
httpx
beautifulsoup4
python-dotenv
sse-starlette
